refactor(ui): Extract space tokens
