+++ packages/ui/src/layout/box/box.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import cx from 'classnames';
 
-import { NO_SPACE, SPACES } from '../../constants';
+import { NO_SPACE, SPACES } from '../../tokens';
 // @ts-ignore
 import css from './box.module.css';
